+++ src/example_agents/document_qa_agent.py
@@ -1,18 +1,29 @@
+from typing import List, Type
+
+from pydantic import Field
 from steamship.agents.functional import FunctionsBasedAgent
 from steamship.agents.llms.openai import ChatOpenAI
-from steamship.agents.mixins.transports.steamship_widget import \
-    SteamshipWidgetTransport
+from steamship.agents.mixins.transports.slack import (
+    SlackTransport,
+    SlackTransportConfig,
+)
+from steamship.agents.mixins.transports.steamship_widget import SteamshipWidgetTransport
+from steamship.agents.mixins.transports.telegram import (
+    TelegramTransport,
+    TelegramTransportConfig,
+)
+from steamship.agents.schema import Tool
 from steamship.agents.service.agent_service import AgentService
 from steamship.agents.tools.question_answering import VectorSearchQATool
+from steamship.invocable import Config
 from steamship.invocable.mixins.blockifier_mixin import BlockifierMixin
 from steamship.invocable.mixins.file_importer_mixin import FileImporterMixin
 from steamship.invocable.mixins.indexer_mixin import IndexerMixin
-from steamship.invocable.mixins.indexer_pipeline_mixin import \
-    IndexerPipelineMixin
+from steamship.invocable.mixins.indexer_pipeline_mixin import IndexerPipelineMixin
 from steamship.utils.repl import AgentREPL
 
 
-class ExampleDocumentQAService(AgentService):
+class DocumentQAAgent(AgentService):
     """DocumentQAService is an example AgentService that exposes:  # noqa: RST201
 
     - A few authenticated endpoints for learning PDF and YouTube documents:
@@ -35,10 +46,53 @@ class ExampleDocumentQAService(AgentService):
         BlockifierMixin,
         IndexerMixin,
         SteamshipWidgetTransport,
+        TelegramTransport,
+        SlackTransport,
     ]
+    """USED_MIXIN_CLASSES tells Steamship what additional HTTP endpoints to register on your AgentService."""
+
+    class DocumentQAAgentConfig(Config):
+        """Pydantic definition of the user-settable Configuration of this Agent."""
+
+        telegram_bot_token: str = Field(
+            "", description="[Optional] Secret token for connecting to Telegram"
+        )
+
+    config: DocumentQAAgentConfig
+    """The configuration block that users who create an instance of this agent will provide."""
+
+    tools: List[Tool]
+    """The list of Tools that this agent is capable of using."""
+
+    @classmethod
+    def config_cls(cls) -> Type[Config]:
+        """Return the Configuration class so that Steamship can auto-generate a web UI upon agent creation time."""
+        return DocumentQAAgent.DocumentQAAgentConfig
 
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
+
+        # Tools Setup
+        # -----------
+
+        # Tools can return text, audio, video, and images. They can store & retrieve information from vector DBs, and
+        # they can be stateful -- using Key-Valued storage and conversation history.
+        #
+        # See https://docs.steamship.com for a full list of supported Tools.
+        self.tools = [VectorSearchQATool()]
+
+        # Agent Setup
+        # ---------------------
+
+        self.set_default_agent(
+            FunctionsBasedAgent(
+                tools=self.tools,
+                llm=ChatOpenAI(self.client),
+            )
+        )
+
+        # Document QA Mixin Setup
+        # -----------------------
 
         # This Mixin provides HTTP endpoints that coordinate the learning of documents.
         #
@@ -50,21 +104,41 @@ class ExampleDocumentQAService(AgentService):
         # That vector index is then available to the question answering tool, below.
         self.add_mixin(IndexerPipelineMixin(self.client, self))
 
-        self._agent = FunctionsBasedAgent(
-            tools=[VectorSearchQATool()],
-            llm=ChatOpenAI(self.client),
-        )
+        # Communication Transport Setup
+        # -----------------------------
 
-        # This Mixin provides HTTP endpoints that connects this agent to a web client
+        # Support Steamship's web client
         self.add_mixin(
             SteamshipWidgetTransport(
-                client=self.client, agent_service=self, agent=self._agent
+                client=self.client,
+                agent_service=self,
+            )
+        )
+
+        # Support Slack
+        self.add_mixin(
+            SlackTransport(
+                client=self.client,
+                config=SlackTransportConfig(),
+                agent_service=self,
+            )
+        )
+
+        # Support Telegram
+        self.add_mixin(
+            TelegramTransport(
+                client=self.client,
+                config=TelegramTransportConfig(
+                    bot_token=self.config.telegram_bot_token
+                ),
+                agent_service=self,
             )
         )
 
 
 if __name__ == "__main__":
-    # AgentREPL provides a mechanism for local execution of an AgentService method.
-    # This is used for simplified debugging as agents and tools are developed and
-    # added.
-    AgentREPL(ExampleDocumentQAService, agent_package_config={}).run()
+    """Run a REPL for this agent.
+
+    The preferred approach, however, is to locate this agent in `api.py` and then run `ship run local`.
+    """
+    AgentREPL(DocumentQAAgent, agent_package_config={}).run()
